Revert "Save hash of expected share from missing audited node (#2025)"
This reverts commit de8070730a71180ce6a6986714e1af0277c93298.

+++ pkg/audit/containment.go
@@ -30,10 +30,10 @@ var (
 type PendingAudit struct {
 	NodeID            storj.NodeID
 	PieceID           storj.PieceID
-	StripeIndex       int64
-	ShareSize         int32
+	StripeIndex       uint32
+	ShareSize         int64
 	ExpectedShareHash []byte
-	ReverifyCount     int32
+	ReverifyCount     uint32
 }
 
 // Containment holds information about pending audits for contained nodes
